requirements.txt
python-dotenv
inquirer
pyfiglet
colorama
termcolor
psycopg
lxml
beautifulsoup4 
tqdm
openai
anthropic
pydantic
google-genai